refactor(quiz): simplify choice handling by integrating state management into ChoicesSection component
